% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study.R
\name{case_study_foa_imat}
\alias{case_study_foa_imat}
\alias{case_study_foa_iotn}
\alias{case_study_foa_itcr}
\title{Utilities for Input of Case Study Funding Opportunity
Announcements}
\usage{
case_study_foa_imat()

case_study_foa_iotn()

case_study_foa_itcr()
}
\description{
\code{case_study_foa_imat()} FOAs for the IMAT program.

\code{case_study_foa_iotn()} FOAs for the IOTN program.

\code{case_study_foa_itcr()} FOAs for the ITCR program.
}
\examples{
case_study_foa_imat()

case_study_foa_iotn()

case_study_foa_itcr()

}
