% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program.R
\name{program_summary}
\alias{program_summary}
\title{Program-level queries and summaries}
\usage{
program_summary(tbl)
}
\arguments{
\item{tbl}{A tibble with column \code{full_foa} giving the FOAs that
define the program.}
}
\value{
\code{program_summary()} returns a tibble ordered by fiscal year
summarizing project activity.
}
\description{
\code{program_summary()} finds all projects, publications,
and citations funded by the Funding Opportunity Announcements
(FOAs) in \code{tbl}, and summarizes by year the number of projects,
total award ammount, publications, citation count, and relative
citation ratio.
}
\details{
\code{program_summary()} can return rows corresponding to years
before and after the years the FOAs were active. Values before
the FOA activity usually represent a grantee assigning credit
to the FOA for a citation published prior to the start of the
FOA, or to a project funded by FOAs not included in the input
\code{tbl} but with the same core project number as a funded
project. Values after the end of the FOA represent publications
that acknowledge the FOA after the termination of the grant.
}
\examples{
foas <- tribble(
    ~full_foa,       ~description,
    "RFA-CA-19-039", "Early-Stage Development of ...",
    "RFA-CA-19-038", "Development of Innovative ...",
    "RFA-CA-19-040", "Advanced Development of ...",
    "RFA-CA-19-041", "Sustained Support..."
)

program_summary(foas)

}
