% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration.R
\name{copublication}
\alias{copublication}
\alias{cocitation}
\title{Program project collaboration through publication and citation}
\usage{
copublication(tbl)

cocitation(tbl)
}
\arguments{
\item{tbl}{a tibble with column \code{full_foa} containing the Funding
Opportunity Announcements defining the projects of interest.}
}
\value{
\code{copublication()} returns a tibble with columns
\itemize{
\item \code{project}: character() project numbers involved in
collaboration
\item \code{n}: integer() total number of publications from the
project.
\item \code{citn}: integer() total number of citations to project
publications.
\item \code{rcr}: numeric() total relative citation rate of project
publications.
\item \code{n_collab_publ}: integer() number of collaborative publications
(publications acknowledging more than one project for support).
\item \code{n_collab}: integer() total number of collaborators
acknowledged.
\item \code{rcr_collab}: numeric() total 'relative citation ratio' for
collaborative publications.
}

\code{cocitation()} returns a tibble with columns
\itemize{
\item \code{project}: character() project number associated with
publications.
\item \code{n}: integer() total number of publications from the
project.
\item \code{citn}: integer() total number of citations to project
publications.
\item \code{rcr}: numeric() total relative citation rate of project
publications.
\item \code{n_self_citn}: integer() number of citations to other
publications of the same project.
\item \code{n_collab_citn}: integer() number of publications cited by other
projects.
\item \code{n_collab}: integer() total number of projects
citing this project.
\item \code{rcr_citn}: numeric() total 'relative citation ratio' of
collaborative cited publications.
}
}
\description{
\code{copublication()} summarizes collaboration between
projects at the time of publication.

\code{cocitation()} summarizes collaborations between
projects through citation.
}
\examples{
foas <- tribble(
    ~full_foa,       ~description,
    "RFA-CA-19-039", "Early-Stage Development of ...",
    "RFA-CA-19-038", "Development of Innovative ...",
    "RFA-CA-19-040", "Advanced Development of ...",
    "RFA-CA-19-041", "Sustained Support of ..."
)

copublication(foas)

cocitation(foas)

}
