% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc_utilities.R
\name{gpc_camelize}
\alias{gpc_camelize}
\alias{gpc_columns_clean}
\alias{gpc_colnames_standardize}
\alias{gpc_datatable}
\alias{gpc_shorten}
\title{Utilities for 'grantpubcite' use in articles}
\usage{
gpc_camelize(x)

gpc_columns_clean(.data)

gpc_colnames_standardize(.data = NULL)

gpc_datatable(.data)

gpc_shorten(x, to = 40L)
}
\arguments{
\item{x}{character(), e.g., \code{project_title} to be truncated.}

\item{.data}{A tibble, usually from an invocation of
\code{reporter_projects()} or \code{icite()}. \code{.data} can also be NULL for
\code{gpc_colnames_standardize()}, in which case the return value is
a tibble summarizing how column names 'from'
\code{reporter_projects()} or \code{icite()} are translated 'to' display
column names.}

\item{to}{integer(1) maximum number of characters to retain in
\code{x}.}
}
\description{
\code{gpc_camelize()} takes an input 'snake_case' vector
and transforms it to CamelCase.

\code{gpc_columns_clean()} is used internally to ensure
consistent column formatting.

\code{gpc_colnames_standardize()} transforms column names
to shorter equivalents, primary useful for (horizontal)
contraction of table displays.

\code{gpc_datatable()} standardizes column names for
consistent display. In non-interactive sessions it retured a
\code{DT::datatable()} for interactive navigation.

\code{gpc_short_title()} shortens titles, e.g., for use as
labels in interactive plots.
}
\details{
\code{gpc_camelize()} is useful in standardizing the user
interface to the API, where the input 'include_fields' are
CamelCase in the API, but the return values are
snake_case. Using \code{gpc_camelize()} (internally) allows the user
to interact with the API using a consistent \code{snake_case} style.

\code{gpc_columns_clean()} performs the following:
\itemize{
\item Remove leading or trailing whitespace on 'character' columns.
\item Coerce columns whose name ends with \verb{_date} to \code{as.Date()}.
\item Coerce 'pmid', 'year' / 'fiscal year' and 'citation_count' to
integer()
\item Remove 'group' annotation from tibbles created during
processing.
}

\code{gpc_datatable()} is a convenience function used in
vignettes to (a) standardize column names for better
formatting; (b) provides a consistent presentation of
DT::datatable without rownames, with pageLength of 5, and
horizontal scrolling, which means that long tables don't
overwhelm articles or spill into margins, and (c) only
transforms the table when not interactive.

Articles typcially use \code{gpc_datatable()} in a code chunk
with \code{include = FALSE}, so that the user sees the datatable but
not code.
}
\examples{
gpc_camelize(c("full_foa", "core_project_num"))

projects <- tribble(
    ~contact_pi_name, ~project_start_date,    ~pmid, ~project_title,
    "Ima Pi ",        "2018-07-15T12:07:00Z", "123", "This is the title",
    "Iman Other ",    "2021-01-14T12:01:00Z", "456", "A short title"
)

clean <-
    projects |>
    gpc_columns_clean()
clean           # no trailing whitespace; dates as Date objects;

gpc_colnames_standardize()      # summarize column name standardization
gpc_colnames_standardize(clean) # shorter column names

clean |>             # standarized colnames;
    gpc_datatable()  # datatable in non-interactive session

clean |>
    mutate(short_title = gpc_shorten(project_title, 13))

}
