% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program.R
\name{program_summary}
\alias{program_summary}
\alias{program_projects}
\title{Program-level queries and summaries}
\usage{
program_summary(tbl)

program_projects(tbl, by = c("full_foa", "core_project_num"))
}
\arguments{
\item{tbl}{A tibble with column \code{full_foa} giving the FOAs that
define the program.}

\item{by}{character(1) indicating how program projects are
summarized -- by \code{full_foa} so that projects funded by more
than one FOA are reported for each FOA, or \code{core_project_num}
so that the summary is by project number..}
}
\value{
\code{program_summary()} returns a tibble ordered by fiscal year
summarizing project activity.
}
\description{
\code{program_summary()} finds all projects, publications,
and citations funded by the Funding Opportunity Announcements
(FOAs) in \code{tbl}, and summarizes by year the number of projects,
total award ammount, publications, citation count, and relative
citation ratio.

\code{program_projects()} returns a tibble of project
numbers and standardized contact PI name and project title.
}
\details{
\code{program_summary()} can return rows corresponding to years
before and after the years the FOAs were active. Values before
the FOA activity usually represent a grantee assigning credit
to the FOA for a citation published prior to the start of the
FOA, or to a project funded by FOAs not included in the input
\code{tbl} but with the same core project number as a funded
project. Values after the end of the FOA represent publications
that acknowledge the FOA after the termination of the grant.

\code{program_projects()} provides a single row for each
project. It chooses as \code{full_foa} the most recent FOA under
which the project was funded. It chooses as \code{contact_pi_name}
and \code{project_title} the name of the contact PI and project
title of the largest award in the most recent year of funding.
}
\examples{
foas <- tribble(
    ~full_foa,       ~description,
    "RFA-CA-19-039", "Early-Stage Development of ...",
    "RFA-CA-19-038", "Development of Innovative ...",
    "RFA-CA-19-040", "Advanced Development of ...",
    "RFA-CA-19-041", "Sustained Support of..."
)

program_summary(foas)

program_projects(foas)

}
