% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program.R
\name{program_summary}
\alias{program_summary}
\alias{program_projects}
\alias{program_publications}
\title{Program-level queries and summaries}
\usage{
program_summary(tbl)

program_projects(tbl, by = c("foa", "project"))

program_publications(tbl)
}
\arguments{
\item{tbl}{A tibble with column \code{full_foa} giving the FOAs that
define the program.}

\item{by}{character(1) indicating how program projects are
summarized -- by \code{foa} so that projects funded by more than one
FOA are reported for each FOA, or \code{project} so that the summary
is by project number across FOA.}
}
\value{
\code{program_summary()} returns a tibble ordered by fiscal year
summarizing project activity. Columns are
\itemize{
\item \code{year}: integer() year of program.
\item \code{project}: integer() number of active projects.
\item \code{amount}: integer() award amount to active projects.
\item \code{publications}: integer() number of publications.
\item \code{cite}: integer() citations to publications in year.
\item \code{rcr}: numeric() sum of relative citation ratios for all
publications in year.
}

\code{program_projects()} returns a tibble summarizing projects
funded under the FOAs. With \code{by = "full_foa"} (default),
columns are:
\itemize{
\item \code{foa}: character() full FOA funding the project.
\item \code{project}: character() core project number.
\item \code{start_date}: date() start date of project
\item \code{end_date}: date() end date of project; maybe in the future
\item \code{contact_pi}: character() name of most-recent contact PI for FOA
and core project number.
\item \code{title}: character() project title.
\item \code{years}: integer() fiscal years of funding; may differ from
project duration (end date - start date.
\item \code{amount}: integer() actual award amount across fiscal years.
}

With \code{by = "project"}, columns are the same but with
\code{start_date}, \code{end_date}, \code{years}, \code{amount} and most recent
contact PI and project title summarized over all FOAs under
which a project was funded.

\code{program_publications()} returns a tibble with columns:
\itemize{
\item \code{foa}: character() full FOA funding the project.
\item \code{project}: character() core project number.
\item \code{pmid}: integer() PubMed identifier.
\item \code{year}, \code{title}, \code{authors}, \code{journal},\code{doi}: publication information.
\item \code{citn}: integer() number of publicatons citing this
publication, from \code{icite()}.
\item \code{rcr}: relative citation ratio, as defined by \code{icite()}.
\item \code{fcr}: field citation rate, as defined by \code{icite()}.
}
}
\description{
\code{program_summary()} finds all projects, publications,
and citations funded by the Funding Opportunity Announcements
(FOAs) in \code{tbl}, and summarizes by year the number of projects,
total award ammount, publications, citation count, and relative
citation ratio.

\code{program_projects()} returns a tibble of project
numbers and standardized contact PI name and project title.

\code{program_publications()} retrieves basic information
for all projects in the program.
}
\details{
\code{program_summary()} can return rows corresponding to years
before and after the years the FOAs were active. Values before
the FOA activity usually represent a grantee assigning credit
to the FOA for a citation published prior to the start of the
FOA, or to a project funded by FOAs not included in the input
\code{tbl} but with the same core project number as a funded
project. Values after the end of the FOA represent publications
that acknowledge the FOA after the termination of the grant.

\code{program_projects()} provides a single row for each
project. It chooses as \code{full_foa} the most recent FOA under
which the project was funded. It chooses as \code{contact_pi_name}
and \code{project_title} the name of the contact PI and project
title of the largest award in the most recent year of funding.
}
\examples{
foas <- tribble(
    ~full_foa,       ~description,
    "RFA-CA-19-039", "Early-Stage Development of ...",
    "RFA-CA-19-038", "Development of Innovative ...",
    "RFA-CA-19-040", "Advanced Development of ...",
    "RFA-CA-19-041", "Sustained Support of..."
)

program_summary(foas)

program_projects(foas)

pubs <- program_publications(foas)

## note that individual publications can be represented by more
## than one project
pubs

## unique publications
unique_pubs <-
    pubs |>
    select(-c("full_foa", "core_project_num")) |>
    distinct()
unique_pubs

## publications per project
pubs |>
    count(core_project_num, sort = TRUE) |>
    left_join(program_projects(foas, by = "project"), by = "core_project_num")

## collaborations; see `?copublication`, `?cocitation`
pubs |>
    count(pmid, sort = TRUE, name = "n_collab") |>
    left_join(unique_pubs, by = "pmid")

}
